+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a configurable option to reveal hidden tracker content before selecting images.
+
 ## [0.2.0] - 2026-08-02
 
 ### Added